feat: wip base layout
