+++ src/App.tsx
@@ -10,7 +10,7 @@ import ProfilePage from "./pages/ProfilePage/ProfilePage";
 import { useAppDispatch, useAppSelector } from "./app/hooks";
 import { isOpenModalValue } from "./features/IsOpenModalSlice";
 import { isOpenProfileValue } from "./features/isOpenProfileModalSlice";
-import Recipes from "./pages/Recipes/Recipes";
+import RecipesPage from "./pages/Recipes/RecipesPage";
 import Categories from "./components/Categories/Categories";
 
 function App() {
@@ -44,7 +44,7 @@ function App() {
         <Route path="sign_in" element={<SignPage signMode="signIn" />} />
         <Route path="all_categories" element={<AllCategoriesPage />}>
           <Route path="" element={<Categories />} />
-          <Route path=":id" element={<Recipes />} />
+          <Route path=":id" element={<RecipesPage />} />
         </Route>
         <Route path="shopping_list" element={<ShoppingListPage />} />
         <Route path="add_recipe" element={<AddRecipePage />} />
