+++ src/pages/AllCategoriesPage/AllCategoriesPage.scss
@@ -13,12 +13,18 @@
 
   &__wrapper {
     display: flex;
-    flex-direction: column;
-    gap: 32px;
     justify-content: flex-start;
     width: 79vw;
     margin-top: 3.7135vh;
     margin-bottom: 6.8965vh;
+  }
+
+  &__wrapper-div {
+    display: flex;
+    flex-direction: column;
+    gap: 32px;
+    justify-content: flex-start;
+    width: 100%;
   }
 
   &__h2 {
