+++ src/pages/Recipes/RecipesPage.scss
@@ -1,0 +1,2 @@
+@import "../../styles/variables";
+@import "../../styles/mixins";
